requirements.txt
APScheduler==3.6.3
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
click==7.1.1
DateTime==4.3
docutils==0.15.2
Flask==1.1.1
flask-googlemaps==0.4.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.0
pandas==1.0.5
pkg-resources==0.0.0
py4j==0.10.9
PyJWT==1.7.1
pymongo==3.10.1
pyspark==3.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.0
WTForms==2.3.1
zope.interface==5.1.0
